requirements.txt
# 核心依赖
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# 向量存储和嵌入
chromadb>=0.4.22
langchain-chroma>=0.1.0
sentence-transformers>=2.2.2
huggingface-hub>=0.20.0

# 文档处理
pypdf>=3.17.0
pdfplumber>=0.10.0
python-docx>=0.8.11
markdown>=3.5.0
unstructured>=0.10.30

# 模型接口
ollama>=0.1.0
openai>=1.12.0
anthropic>=0.25.0

# API框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# 前端框架
streamlit>=1.29.0
gradio>=4.0.0

# 工具和工具包
langchain-experimental>=0.0.50
langgraph>=0.0.20

# 工具和工具包
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0

# 配置和日志
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0